pyproject.toml
[tool.poetry]
name            = "oleg"
version = "0.1.3"
description     = "A service bot to provide the server with all the features it needs."
authors         = ["stefanlight <steffmukhin@gmail.com>"]
license         = "MIT License"
readme          = "README.md"

[tool.poetry.dependencies]
python          = ">=3.12,<3.13"
cachetools = "^5.3.2"

hikari          = "^2.0.0.dev122"
hikari-miru     = "^3.4.0"
hikari-crescent = "^0.6.6"

orjson          = "^3.9.10"
uvloop          = "^0.19.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.2.2"

[build-system]
requires        = ["poetry-core"]
build-backend   = "poetry.core.masonry.api"
